requirements.txt
aioodbc
discord.py
sanic
